requirements.txt
streamlit==1.53.1
langchain==1.2.8
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-openai==1.1.7
mem0ai==1.0.1
python-dotenv==1.2.1
